+++ src/component/activity_list/chart.rs
@@ -1,4 +1,4 @@
-use chrono::{NaiveDateTime};
+use chrono::NaiveDateTime;
 use tui::{
     backend::Backend,
     layout::Constraint,
@@ -10,13 +10,10 @@ use tui::{
 };
 
 use crate::{
-    app::{App},
-    event::{
-        keymap::{MappedKey},
-    },
+    app::App,
+    event::keymap::MappedKey,
+    store::activity::{SortBy, SortOrder},
 };
-
-
 
 pub fn handle(_app: &mut App, _key: MappedKey) {}
 pub fn draw<B: Backend>(
@@ -24,7 +21,9 @@ pub fn draw<B: Backend>(
     f: &mut Frame<B>,
     area: tui::layout::Rect,
 ) -> Result<(), anyhow::Error> {
-    let activities = &app.unsorted_filtered_activities();
+    let activities = &app
+        .activities()
+        .sort(&SortBy::Date, &SortOrder::Asc);
     let times: Vec<i64> = activities.timestamps();
     let paces: Vec<i64> = activities.meter_per_hours();
     let tmax = times.iter().max();
@@ -41,7 +40,8 @@ pub fn draw<B: Backend>(
     }
     let pmin = pmin.unwrap();
     let pmax = pmax.unwrap();
-    let data: Vec<(f64, f64)> = activities.to_vec()
+    let data: Vec<(f64, f64)> = activities
+        .to_vec()
         .iter()
         .map(|a| {
             let ts = a.start_date.unwrap().timestamp();
@@ -50,11 +50,11 @@ pub fn draw<B: Backend>(
         .collect();
     let mut current = vec![];
     if let Some(selected) = app.activity_list.table_state().selected() {
-        let activities = app.filtered_activities();
+        let activities = app.activities();
         if let Some(a) = activities.get(selected) {
-            if let Some(a) = app.activities.find(a.id) {
-                    current.push((a.start_date.unwrap().timestamp() as f64, *pmin as f64));
-                    current.push((a.start_date.unwrap().timestamp() as f64, *pmax as f64));
+            if let Some(a) = activities.find(a.id) {
+                current.push((a.start_date.unwrap().timestamp() as f64, *pmin as f64));
+                current.push((a.start_date.unwrap().timestamp() as f64, *pmax as f64));
             }
         }
     }
@@ -71,17 +71,17 @@ pub fn draw<B: Backend>(
             .marker(Marker::Braille)
             .graph_type(GraphType::Scatter)
             .style(Style::default().fg(Color::Magenta)),
-        Dataset::default().data(&current)
+        Dataset::default()
+            .data(&current)
             .name("Selected")
             .marker(Marker::Braille)
             .graph_type(GraphType::Line)
             .style(Style::default().fg(Color::Green)),
     ];
     let yaxisstep = (pdiff as f64 / area.height as f64) as usize;
-    let yaxis =
-        (*pmin..*pmax).step_by(if yaxisstep > 0 { yaxisstep } else { 1 });
+    let yaxis = (*pmin..*pmax).step_by(if yaxisstep > 0 { yaxisstep } else { 1 });
     let xaxisstep = (tdiff as f64 / 5.0) as usize;
-    let xaxis = (*tmin.unwrap()..*tmax.unwrap()).step_by(if xaxisstep > 0 { xaxisstep } else {1});
+    let xaxis = (*tmin.unwrap()..*tmax.unwrap()).step_by(if xaxisstep > 0 { xaxisstep } else { 1 });
     let chart = Chart::new(datasets)
         .hidden_legend_constraints((Constraint::Max(1), Constraint::Max(1)))
         .block(Block::default().borders(Borders::all()))
@@ -105,7 +105,10 @@ pub fn draw<B: Backend>(
             Axis::default()
                 .title(Span::styled("Pace", Style::default().fg(Color::Red)))
                 .style(Style::default().fg(Color::White))
-                .bounds([*pmin as f64, *pmax as f64 + (pdiff as f64 / activities.len() as f64)])
+                .bounds([
+                    *pmin as f64,
+                    *pmax as f64 + (pdiff as f64 / activities.len() as f64),
+                ])
                 .labels(
                     yaxis
                         .map(|p| Span::from(app.unit_formatter.pace(3600, p as f64)))