+++ CHANGELOG.md
@@ -3,4 +3,7 @@ CHANGELOG
 
 ## master
 
+Features:
+
 - Highlight and allow selection of current split in activity view.
+- Added activity ranking.